Moves monitor delay calculation out of trials, since it seems it might be needed in other places
