+++ allensdk/brain_observatory/behavior/data_objects/running_speed/running_speed.py
@@ -1,5 +1,3 @@
-
-import json
 from typing import Optional
 
 import pandas as pd
